GH-7785 Checkpoint. Moving guest checkout functionality to new login-register component
