+++ projects/storefrontlib/src/cms-components/user/login-form/login-form.component.spec.ts
@@ -1,8 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
-import { By } from '@angular/platform-browser';
-import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import {
   AuthRedirectService,
@@ -13,7 +11,6 @@ import {
   WindowRef,
 } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
-import { CheckoutConfigService } from '../../checkout';
 import { LoginFormComponent } from './login-form.component';
 import createSpy = jasmine.createSpy;
 import { FormErrorsModule } from '../../../shared/index';
@@ -37,20 +34,6 @@ class MockRedirectAfterAuthService {
 }
 class MockGlobalMessageService {
   remove = createSpy();
-}
-
-class MockActivatedRoute {
-  snapshot = {
-    queryParams: {
-      forced: false,
-    },
-  };
-}
-
-class MockCheckoutConfigService {
-  isGuestCheckout() {
-    return false;
-  }
 }
 
 describe('LoginFormComponent', () => {
@@ -78,8 +61,6 @@ describe('LoginFormComponent', () => {
           useClass: MockRedirectAfterAuthService,
         },
         { provide: GlobalMessageService, useClass: MockGlobalMessageService },
-        { provide: ActivatedRoute, useClass: MockActivatedRoute },
-        { provide: CheckoutConfigService, useClass: MockCheckoutConfigService },
       ],
     }).compileComponents();
   }));
@@ -160,31 +141,6 @@ describe('LoginFormComponent', () => {
         email_lowercase,
         password
       );
-    });
-  });
-
-  describe('guest checkout', () => {
-    it('should show "Register" when forced flag is false', () => {
-      const registerLinkElement: HTMLElement = fixture.debugElement.query(
-        By.css('.btn-register')
-      ).nativeElement;
-      const guestLink = fixture.debugElement.query(By.css('.btn-guest'));
-
-      expect(guestLink).toBeFalsy();
-      expect(registerLinkElement).toBeTruthy();
-    });
-
-    it('should show "Guest checkout" when forced flag is true', () => {
-      component.loginAsGuest = true;
-      fixture.detectChanges();
-
-      const guestLinkElement: HTMLElement = fixture.debugElement.query(
-        By.css('.btn-guest')
-      ).nativeElement;
-      const registerLink = fixture.debugElement.query(By.css('.btn-register'));
-
-      expect(registerLink).toBeFalsy();
-      expect(guestLinkElement).toBeTruthy();
     });
   });
 });
