+++ projects/storefrontlib/src/cms-components/user/login-form/login-form.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { ActivatedRoute } from '@angular/router';
 import {
   AuthRedirectService,
   AuthService,
@@ -9,8 +8,9 @@ import {
   WindowRef,
 } from '@spartacus/core';
 import { Subscription } from 'rxjs';
-import { CheckoutConfigService } from '../../checkout/services/checkout-config.service';
 import { CustomFormValidators } from '../../../shared/index';
+import { ActivatedRoute } from '@angular/router';
+import { CheckoutConfigService } from '../../checkout/services';
 
 @Component({
   selector: 'cx-login-form',
@@ -19,16 +19,27 @@ import { CustomFormValidators } from '../../../shared/index';
 export class LoginFormComponent implements OnInit, OnDestroy {
   sub: Subscription;
   loginForm: FormGroup;
-  loginAsGuest = false;
+
+  /**
+   * @deprecated since version=2.1
+   * Use constructor (auth: AuthService, globalMessageService: GlobalMessageService, fb: FormBuilder, authRedirectService: AuthRedirectService, winRef: WindowRef)
+   */
+  constructor(
+    auth: AuthService,
+    globalMessageService: GlobalMessageService,
+    fb: FormBuilder,
+    authRedirectService: AuthRedirectService,
+    winRef: WindowRef,
+    checkoutConfigService: CheckoutConfigService,
+    activatedRoute: ActivatedRoute
+  );
 
   constructor(
     protected auth: AuthService,
     protected globalMessageService: GlobalMessageService,
     protected fb: FormBuilder,
     protected authRedirectService: AuthRedirectService,
-    protected winRef: WindowRef,
-    protected activatedRoute: ActivatedRoute,
-    protected checkoutConfigService: CheckoutConfigService
+    protected winRef: WindowRef
   ) {}
 
   ngOnInit(): void {
@@ -42,12 +53,6 @@ export class LoginFormComponent implements OnInit, OnDestroy {
       ],
       password: ['', Validators.required],
     });
-
-    if (this.checkoutConfigService.isGuestCheckout()) {
-      this.loginAsGuest = this.activatedRoute?.snapshot?.queryParams?.[
-        'forced'
-      ];
-    }
   }
 
   submitForm(): void {
